backend/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
fuzzywuzzy==0.18.0
google-api-core==2.11.0
google-auth==2.17.1
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
googleapis-common-protos==1.59.0
gptcache==0.1.11
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
langchain==0.0.140
Levenshtein==0.21.0
llama-index==0.5.15
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
PyJWT==2.7.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-Levenshtein==0.21.0
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.0.0
regex==2023.3.23
requests==2.28.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.4.47
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
tqdm==4.65.0
twilio==8.2.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
